refactor: CR changes, adapted to WordSearcher
