+++ parser_columns/marcaine_medicine_parser.py
@@ -1,16 +1,20 @@
-from parser_columns.base_parser import BaseParser
-from utils.find_similar_words import find_similar_words
+import pandas as pd
+
 from constants import POSITIVE_VALUE, NEGATIVE_VALUE, MISSING_VALUE
+from parser_columns.base_parser import BaseParser
+from utils.word_searcher import WordSearcher
 
 MARCAINE = "מרקאין"
 
-class DoesContainMarcaineMedicineParser(BaseParser):
+
+class MarcaineMedicineParser(BaseParser):
     def __init__(self) -> None:
         super().__init__()
+        self.word_searcher = WordSearcher(word=MARCAINE)
 
     def parse_document(self, document_content: str) -> str:
         try:
-            similar_words_in_text = find_similar_words(document_content, MARCAINE)
+            similar_words_in_text = self.word_searcher.find_similar_words(document_content)
             if len(similar_words_in_text) > 0:
                 return POSITIVE_VALUE
             else:
